feat:remove mecab and add setup.py with setuptools
